cmd/gateway/s3: support a MINIO_GATEWAY_DEFAULT_REGION env var
With scaleway the `fr-par` region isn't detected correctly and the
default region is `us-east-1` which doesn't work with scaleway.
This renders the gateway unsuable (403 responses with wrong region
`us-east-1`).
This commit adds a new env var `MINIO_GATEWAY_DEFAULT_REGION` that
allows to provide a default region fallback when no region is detected
from the storage url.

It's intended to be as light a fork as possible to be able to easily
keep up with upstream changes. Especially since gateway will undergo
significant changes in the comming months.